requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ninja==0.16.1
pydantic==1.8.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.0
PyPDF2==1.26.0
reportlab==3.5.66
Pillow==8.1.0
django-samurai==0.1.1
bleach==4.1.0
